requirements.txt
nltk>=3.4.5
tweepy==3.7.0
Unidecode==1.0.23
numpy==1.16.1
tmdbv3api==1.3.2
requests==2.21.0
pandas==0.24.1
Keras==2.2.4
Pillow==6.0.0
google_api_python_client==1.7.8
scikit_learn==0.20.3
